src/analytics-worker-task/requirements.txt
motor==3.3.2
pymongo==4.6.0
kafka-python==2.0.2
pydantic==2.5.0
structlog==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1